requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-RESTful==0.3.10
Flask-Bcrypt==1.0.1
SQLAlchemy-serializer==1.4.1
python-dotenv==1.0.0
gunicorn==21.2.0